+++ source/fortran/numerics.f90
@@ -100,9 +100,9 @@ module numerics
   public
 
   !+ad_vars  ipnvars FIX : total number of variables available for iteration
-  integer, parameter :: ipnvars = 148
+  integer, parameter :: ipnvars = 149
   !+ad_vars  ipeqns  FIX : number of constraint equations available
-  integer, parameter :: ipeqns = 78
+  integer, parameter :: ipeqns = 79
   !+ad_vars  ipnfoms FIX : number of available figures of merit
   integer, parameter :: ipnfoms = 18
 
@@ -351,8 +351,10 @@ module numerics
        'Eich critical separatrix density ',   &
        !+ad_varc  <LI> (77) TF coil current per turn upper limit 
        'TFC current per turn upper limit ',    &
-       !+ad_varc  <LI> (78) Reinke criterion impurity fraction lower limit (itv  147 freinke)</UL>
-       'Reinke criterion fZ lower limit  '    &
+       !+ad_varc  <LI> (78) Reinke criterion impurity fraction lower limit (itv  147 freinke)
+       'Reinke criterion fZ lower limit  ',   &
+       !+ad_varc  <LI> (79) F-value for max peak CS field (itv  149 fbmaxcs)</UL>
+       'Peak CS field upper limit        '    &
        /)
        ! Please note: All strings between '...' above must be exactly 33 chars long
        ! Each line of code has a comma before the ampersand, except the last one.
@@ -665,10 +667,12 @@ module numerics
        'fgwped        ', &
        !+ad_varc  <LI> (146) fcpttf : F-value for TF coil current per turn limit (constraint equation 77)
        'fcpttf        ', &
-       !+ad_varc  <LI> (147) freinke : F-value for Reinke detachment criterion (constraint equation 78)</UL>
+       !+ad_varc  <LI> (147) freinke : F-value for Reinke detachment criterion (constraint equation 78)
        'freinke       ', &
-       !+ad_varc  <LI> (148) fzactual : fraction of impurity at SOL with Reinke detachment criterion</UL>
-       'fzactual      ' &
+       !+ad_varc  <LI> (148) fzactual : fraction of impurity at SOL with Reinke detachment criterion
+       'fzactual      ', &
+       !+ad_varc  <LI> (149) fbmaxcs : F-value for max peak CS field (con. 79, itvar 149)</UL>
+       'fbmaxcs       ' &
        /)
 
   character(len=14), dimension(:), allocatable :: name_xc
@@ -834,7 +838,8 @@ module numerics
        0.500D0, &  !  145
        0.001D0, &  !  146
        0.001D0, &  !  147
-       1.00D-8  &  !  148
+       1.00D-8, &  !  148
+       0.001D0  &  !  149
        /)
 
   !+ad_vars  boundu(ipnvars) /../ : upper bounds used on ixc variables during
@@ -987,7 +992,8 @@ module numerics
        1.000D0, &  !  145
        1.000D0, &  !  146
        1.000D0, &  !  147
-       1.000D0  &  !  148
+       1.000D0, &  !  148
+       1.000D0  &  !  149
        /)
 
   real(kind(1.0D0)), dimension(ipnvars) :: bondl = 0.0D0